Make state manager url hot-reloadable
This could solve the issue of the API having to re-template and restart
each time the state manager is deployed, due to the fact that Nomad
selects a random port for the state manager.

Previously this has been solved by using the public URL however this
results in going through Cloudflare and Caddy again.

I could also put the state manager on a static port, but I'm not sure
whether Consul would still re-template in the 5 second window that the
service is deleted. This is also recommended against in the docs as it
makes it harder for Nomad to schedule jobs.