Update PhotonStateManager related configuration (#865)
- Adds support for the
[key-per-file](https://learn.microsoft.com/en-us/dotnet/core/extensions/configuration-providers#key-per-file-configuration-provider)
configuration provider to the app.
- Also fixes the lambda for adding `IPhotonStateApi` to make it pull an
updated value for the base URL.
- Miscellaneous refactoring to PhotonStateManager config, including
better use of DI and `IConfiguratioon`.

This includes several breaking changes for PhotonStateManager:
- `PHOTON_TOKEN` environment variable -> `PhotonOptions__Token`
- `ENABLE_HTTPS` environment variable no longer used, consult ASP.NET
docs for alternative

---

This is primarily intended to solve the issue of the API having to
re-template and restart each time the state manager is deployed, due to
the fact that Nomad selects a random port for the state manager.
Currently the URL is stored as an env variable which requires a
container restart to reflect a change.

Previously this has been solved by using the public URL however this
results in going through Cloudflare and Caddy again.

I could also put the state manager on a static port, but I'm not sure
whether Consul would still re-template in the 5 second window that the
service is deleted. This is also recommended against in the docs as it
makes it harder for Nomad to schedule jobs.