Initialize Mercurial Gauntlet on story skip
Seeing errors in the logs about people attempting to access Mercurial
Gauntlet without the required DbQuestWalls initialized. It seems the
gap here is story skip - you skip past the MG story but without having
the data initialized. Attempting to read the story after this will not
send a request to the server.

Will need to run a DB query after this to fix affected saves.

Plus a usual bunch of misc refactoring and global query filter usage.