Fix failure to get RoomId in OnCloseGame callback
There are some errors in the logs where the OnCloseGame callback of the
StateManagerPlugin is failing because there is no RoomId property.

This appears to happen if a player joins a game and it closes instantly,
typically due to connection issues. It seems in this case the code in
OnCreateGame which would initialize the RoomId does not have a chance to
run.

These headers are not required for the state manager or for Zena, and
are only used for recording time attack clears in GameLogicPlugin. So
the simplest thing to do is remove them.