add website tests
This doesn't really work, there's a lot to consider with regards to how
you communicate with the API. Using a real API in a CI scenario is never
going to work. You would need to pre-seed with a lot of data to make it
useful. MSW sounded promising but seems to have issues with sveltekit
preloading bypassing the handlers (may be fixable).