Add MasterAsset extensions (#928)
Adds support for specifying an extension, with an optional feature flag,
to MasterAsset. This will generate code that loads an additional file to
override the contents of a master asset at load time, which will be
gated by the feature flag if specified.

This is primarily to enable some small-scale "mods" that I want to
implement, specifically around boosting daily endeavour rewards in the
absence of regular events. All such mods will be put behind feature
flags so that they can be turned off and this project can still act as a
'reference implementation' of a vanilla Dragalia Lost backend.

A secondary purpose, which has been used as a test run here, is to
'patch' the MasterAsset where it contains incorrect data. A previous
workaround for welfare adventurer story IDs being incorrect has been
repurposed to use this new functionality.