FIx result code being null if returning an object directly
Modern controllers use DragaliaResult<TData> which returns an
IActionResult by implementing IConvertToActionResult. This bypasses the
mechanism implemented in #1016 inside DragaliaControllerBase which sets
the ResultCode based on the use of Ok() or Code().

To have the result code set in these cases, use an action filter that
runs after all actions and attempts to determine if the result is a
DragaliaResult. I've had to implement an abstract class for this as it
wouldn't be possible to pass a value for the generic argument.

This mechanism replaces any need to set the result code manually in
DragaliaControllerBase.

Also re-ordered the constructors of DragaliaResult to make them a tad
more consistent.