Fix co-op softlock for player 2 (#1030)
Incoming /dungeon_record/record_multi requests were being subjected to
output caching after #1016, because the Photon server does not set
Request-Token or SID so the vary-by headers were always the same. This
lead to P2 being served P1's response which lead to a softlock.

Fix this by only caching requests where a Request-Token was present in
the first place. This also replaces the existing workaround to disable
output caching in the automated tests.