Add basic save editor to website (#1036)
Add a page where you can add things to your gift box on the website.

Should allow users to access inaccessible entities e.g. raid welfare
characters, without the save editor. Also opens up the possibility of
save editing things previously out of reach due to not being in the
exported save, such as Kaleidoscape points.

This is the frontend only, backend is still to be done. As such it is
not visible on the home page and is locked behind a flag.