Use batch reward handler to claim presents
On testing the save editor I was hitting a fail point I added when
attempting to grant multiple dragons without using the batch handler.

We need to use the batch handler for presents to ensure all types of
presents can be claimed - I suspect multiple dragons in one present was
not possible before the editor.