Don't cache requests without a SID header
Prevents unauthenticated requests becoming cached for all users. The
Request-Token is not as random as we thought. It is calculated by

`(CurrentRequestId + 1) & 0xFFFFFF | (UnixTime(DateTime.UtcNow) << 24)`

which could mean that two clients making a series of requests at the
same time could receive the same token, i.e. it is not actually pseudo-
random as originally assumed.