Fix server error when claiming amber before initializing kaleidoscape
Currently the server throws an error if you try and claim a present or
other reward containing Kaleidoscape amber before unlocking the
Kaleidoscape mode. This is because it can't find the DbPlayerDmodeInfo
to update.

This was never really a problem until the introduction of the save
editor, where people seem to be quite commonly hitting this case.

Introduce a reward handler for these which can initialize the
Kaleidoscape data when attempting to claim a present.

TODO:
- [ ] Check the behaviour for claiming a dmode present and THEN entering
kaleidoscape for the first time. If the game still calls /dmode/entry
we will get a PK conflict