Fix server error when claiming amber before initializing kaleidoscape (#1066)
Currently the server throws an error if you try and claim a present
orother reward containing Kaleidoscape amber before unlocking the
Kaleidoscape mode. This is because it can't find the DbPlayerDmodeInfo
to update.

This was never really a problem until the introduction of the save
editor, where people seem to be quite commonly hitting this case.

Introduce a reward handler for these which can initialize the
Kaleidoscape data when attempting to claim a present.

It seems that having any kaleidoscape data initialized will then lead
the game to skip the kaleidoscape intro (and calling /dmode/entry) so
there shouldn't be any PK conflicts.