Optimize background image
Split the background images into wide and narrow variants so that we
don't end up loading the full image where half of it will not be seen
because of screen size.

The previous images were 384 KB for light mode, and 229 KB for dark
mode. The new images are 145 KB / 166 KB for narrow/wide in light mode,
and 83.8 KB / 104 KB for narrow/wide in dark mode. So savings of about
50%.

Also looked into using <enhanced:img> but think this is a non-starter
if we want to load different images based on CSS media queries