Exclude completed buildings from working_carpenter_num
Contributes to #436. Changes the used carpenter count calculation to be based on the number of buildings currently under construction which aren't finished, rather than any building that has a non-zero BuildEndDate.

This is because if you start an upgrade that takes e.g. 10 seconds, we report you have 3/4 builders. That's fine. But when the client sees that build complete using its internal timer, it then believes 4/4 builders are available, but we currently would say it's still 3/4 until you actually call levelup_end by tapping the building. This desync was causing the FortBuildCarpenterBusy exception.