Parallelize integration tests (#1170)
Use AssemblyFixture to run the integration tests in parallel, rather than in series with a database reset in-between. This allows the tests to complete a fair bit faster.  
Dawnshard is effectively a multi-tenant system, so in theory each test can run in isolation under its own viewer ID without affecting others. The only exceptions are things that are public information, like the news data on the website and time attack rankings.

- Don't set AssertionOptions globally as this appears to trigger exceptions inside FluentAssertions if done from multiple threads
- Various changes to make the tests not interfere with each other, typically by scoping queries and updates in tests to the current viewer ID