Fix login bugs (#1212)
- Fix repeated "can't get challenge string" errors - I suspect this is
bots or people otherwise directly browsing to /ouath. Instead, return a
plain bad request if visiting with wrong parameters. Check
`session_token_code` query param first as this should _always_ be in the
URL if coming from BaaS, if we no longer see the missing cookie errors
after this it's likely it was from direct requests.
- Store the original page when redirecting to /unauthorized and use it
when logging in - this will allow direct linking to e.g.
/account/save-editor
- Misc fixes to `Card.Title` to avoid including a heading inside a
heading