Fix DbUpdateException when story skipping
Fix an exception from PK conflicts on the MG missions, if you story skip after unlocking MG.

In general InitializeWall should be idempotent, but it will always try and add the missions regardless. Fix this by assuming that the missions are not initialized if the wall entities are initialized.

The V18Update handles adding wall missions from a save import POV where the wall levels are initialized but the missions aren't.