Expand CSP policy (#1230)
Adding `object-src: none` and `base-uri: none` as per the guidance of
https://csp-evaluator.withgoogle.com/


![image](https://github.com/user-attachments/assets/fc44cfe9-beb3-4353-bf8a-f258d5dd0e02)

Note that SvelteKit does not support unsafe-inline fallbacks; we would
need to add `strict-dynamic` which is officially unsupported at the
minute. Adding `unsafe-inline` also removes the nonce and breaks the
entire page as the browser ignores it due to the unsafe-hashes
workaround, but the missing nonce means no other scripts can execute.

`require-trusted-types` also is apparently not supported.