Add User-Agent to CSP reports (#1253)
We're getting a lot of CSP violation reports that I am unable to
reproduce. The `unsafe-hashes` part of the CSP is only supported by CSP
Level 3 browsers, so I'm wondering if the reports are coming from old
browsers.

If I can't see a clear pattern in the reports and figure out a way to
reproduce them, I will probably just turn them off, as I'm confident the
website still works as intended (in my environment/browser at least).