Force light mode for webview pages (#1266)
Don't inject the mode watcher script for webview pages, otherwise users
who have set a preference on the main website will see dark mode pages
in-game, which contrasts with the light mode theme of the other game UI
elements.

This only seems to happen on iOS, possibly because the browser local
storage is shared between web-view and the system browser?