Enforce friend limit (#1292)
Enforce the limit to friends, which ranges between 25 and 175 depending
on player level.

Some complexity around requests - we want to check 'potential' friends
as well, i.e. incoming and outgoing requests, so that it isn't possible
to exceed the limit by having many requests accepted at once. The game
has a handful of status codes to communicate this for both sides of the
interaction.