Add settings menu to website (#1314)
Adds a basic settings menu to the 'Profile' page of the website for
logged in users.

This only has one setting for now, around the receipt of the daily
material gifts. I'm adding this now because I anticipate implementing
the 'proper' helper system, which is less user-friendly than the current
static helper lists. At that time, I'd like to offer a way to change
back to the legacy behaviour, and the website seems the best place to do
it.