Fix disabling settings
When a switch is toggled off, it is absent from the form data. This leads to it being defaulted to true, meaning it is impossible to toggle off. Iterate a known set of setting keys so that our request object contains all of them with true/false values. Additionally introduce validation that all keys are in the request object with JsonRequired.