Implement real helper system (#1322)
We currently show a static list of pre-defined meta helpers that can be
used an infinite number of times. This PR introduces the vanilla helper
system, in which you must choose from a list of your friends or from
random users. If you use a helper, it then becomes unavailable until the
associated player clears a quest, or the next reset passes.

This is arguably a regression in terms of quality-of-life, so an opt-out
is provided on the website to change back to the old behaviour.

TODO:

- [x] squash migrations