Block sending of duplicate friend requests earlier on
We currently return an error for sending a friend request to an existing friend by catching an FK exception from the DB. This works, but yields two error logs each time it happens which leads to false alarm when looking at the logs.

We can instead return `FriendApplyExists` from `id_search` after a manual query to prevent this, but leave the FK check just in case this is bypassed somehow.