Event drops refactor (#1359)
Closes #489

Adds the event drops before a quest starts, so that they show as drop
indicators in the quest. Also simplifies drop handling as they are now
subject to the same process of summing up the defeated enemies and
applying their drops.

Also implements a visual fix to add the bonus T1 materials with a
particular `factor` to show as bonus drops. Stops bonuses from applying
to T2 and T3 materials.