User impersonation page (#1373)
Adds an admin-only page that allows access to the existing user
impersonation functionality. Previously this had to be accessed via the
GraphQL API which was highly inconvenient and unintuitive.

The page is deliberately quite spartan as it is designed only for admins
/ devs to use.

Access to it and the associated endpoints is driven by the presently
unused 'IsAdmin' column in the Players table. Server administrators must
manually edit rows in the database to mark them as admins and allow them
to access this page.