Fix softlock when exceeding maximum EXP (#1374)
A player reported that every quest clear was soft-locking his client on
a blank screen. On further investigation, the game was throwing an
exception:

```
08-09 00:09:51.350 28554 28600 E Unity   : NullReferenceException: Object reference not set to an instance of an object.
08-09 00:09:51.350 28554 28600 E Unity   :   at Gluon.QuestResultModel+PlayerLevelupModel.LoadData () [0x00000] in <00000000000000000000000000000000>:0 
08-09 00:09:51.350 28554 28600 E Unity   :   at Gluon.QuestResultModel.LoadData () [0x00000] in <00000000000000000000000000000000>:0 
08-09 00:09:51.350 28554 28600 E Unity   :   at Gluon.QuestResultTopPage.InitByData () [0x00000] in <00000000000000000000000000000000>:0 
08-09 00:09:51.350 28554 28600 E Unity   :   at Gluon.QuestResultScene+<OnEnterSceneCoroutine>d__47.MoveNext () [0x00000] in <00000000000000000000000000000000>:0 
08-09 00:09:51.350 28554 28600 E Unity   :   at UnityEngine.SetupCoroutine.InvokeMoveNext (System.Collections.IEnumerator enumerator, System.IntPtr returnValueAddress) [0x00000] in <00000000000000000000000000000000>:0 
```

This suggested the game was trying to show a level up screen where one
could not be loaded.

The player had already reached level 250, and the server will not allow
levelling up past 250 because there is no data for it:


https://github.com/SapiensAnatis/Dawnshard/blob/eed6fbb16889b60862bfabbb61dac76ad625aee5/DragaliaAPI/DragaliaAPI/Features/Player/UserService.cs#L206-L210

but we will still grant experience. It seems that if you gain a certain
amount of experience past the level 250 amount, the client tries to
trigger a level-up and fails. It's possible that it has its own
exponential curve calculation of some kind that leads it to this
behaviour.

The fix here is to never increment a user's EXP beyond the total value
for level 250. If we return that 0 EXP was granted and the user is level
250, the issue is fixed.