Helper fixes (#1381)
The following scenario appears to have occurred:

1. Player A sends a friend request to Player B
2. Player A imports an invalid save that crashes on a save update in
/load/index due to an invalid dragon. The save import has deleted their
helper row but no new row can be created since they can't enter the game
3. Player B now cannot view their friend requests as it crashes on
trying to map the helper

To prevent this from happening, the creation of a helper has been added
to the save import process itself.

Additionally, since the 'Player A' scenario where a user locks
themselves out of the game appears to keep happening, basic validation
has been implemented to reject saves outright if they contain bad IDs.
This likely incurs a performance penalty but we want to discourage the
use of manual save imports in favour of the website tools anyway.